add clear shoppping car function

+++ MeiQianZheng/Tools/CommodityManageTool.m
@@ -2,7 +2,7 @@
 //  CommodityManageTool.m
 //  ThoughtWorks
 //
-//  Created by msn on 16/5/25.
+//  Created by msn on 16/7/17.
 //  Copyright © 2016年 ZYL. All rights reserved.
 //
 
@@ -24,63 +24,28 @@ static FMDatabase * _db;
          [fileMgr removeItemAtPath:path error:nil];;
     };
     _db = [FMDatabase databaseWithPath:path];
-    
     [_db open];
-    
     // 2.创建商品信息表
     [_db executeUpdate:@"CREATE TABLE IF NOT EXISTS t_commodity_list (id integer PRIMARY KEY, barcode string NOT NULL , name string,unit string,categoryid string,category string,subcategory string,price real,discountype string);"];
-    
     //3.创建商品种类表
     [_db executeUpdate:@"CREATE TABLE IF NOT EXISTS t_category (categoryid integer PRIMARY KEY,category string unique,isExistInShoppingCar integer)"];
-    
     //4.创建优惠商品表(单品满100减10块)
     [_db executeUpdate:@"CREATE TABLE IF NOT EXISTS t_discount_3 (id integer PRIMARY KEY, barcode string NOT NULL)"];
-    
     //5.创建购物车商品列表
     [_db executeUpdate:@"CREATE TABLE IF NOT EXISTS t_shoppingcar (id integer PRIMARY KEY, barcode string , name string,unit string,categoryid string,category string,subcategory string,price real,discountype string,count integer);"];
-    
     //6.创建购物车商品种类表
     [_db executeUpdate:@"CREATE TABLE IF NOT EXISTS t_shoppingcar_category (categoryid integer PRIMARY KEY,category string unique)"];
 }
 
-/*
-  添加/更新 优惠商品信息
- */
-+ (void) addDiscount_3:(NSArray *)barcodeArray {
-    //先清空原来的优惠集合
-    [_db executeUpdate:@"DELETE FROM t_discount_3 "];
-    //再将新集合插入
-    for (CommodityModel * item in barcodeArray) {
-         [_db executeUpdateWithFormat:@"INSERT INTO t_discount_3 (barcode) VALUES (%@);",item.barcode];
+#pragma mark --向商品列表中添加商品
++(BOOL)addCommodityInList:(CommodityModel *)item{
+    if (item == nil) {
+        return false;
     }
-}
-
-/*
- * 判断某个商品是否满足优惠
- */
-+ (BOOL)isDiscountCommodity:(CommodityModel *)item{
-    // 得到结果集
-    FMResultSet * set = [_db executeQuery:@"SELECT * FROM t_discount_3;"];
-    // 不断往下取数据
-    while (set.next) {
-        // 获得当前所指向的数据
-        NSString * barcode = [set stringForColumn:@"barcode"];
-        if ([barcode isEqualToString:item.barcode]) {
-            return YES;
-        }
-    }
-    return NO;
-}
-
-+(void)addCommodityInList:(CommodityModel *)item{
-    
     FMResultSet * set = [_db executeQueryWithFormat:@"SELECT * FROM t_commodity_list WHERE barcode = %@",item.barcode];
-    
     if (set.next) {
         [_db executeUpdateWithFormat:@"DELETE FROM t_commodity_list WHERE barcode = %@",item.barcode];
     }
-    
-    //到t_category中查询，看是否已存在该category，如果没有则插入
     NSInteger categoryId = [CommodityManageTool queryCategoryTypeWithCategoryName:item.category];
     if (categoryId == -1) {
        [_db executeUpdateWithFormat:@"INSERT INTO t_category (category) VALUES (%@);",item.category];
@@ -94,17 +59,14 @@ static FMDatabase * _db;
             promotionType = @"salesAll";
         }
     }
-    
-    
     [_db executeUpdateWithFormat:@"INSERT INTO t_commodity_list (barcode,name,unit,categoryid,category,subcategory,price,discountype) VALUES (%@,%@,%@,%ld,%@,%@,%f,%@);",item.barcode,item.name,item.unit,(long)categoryId,item.category,item.subCategory,item.price,promotionType];
+    return true;
 }
 
-+ (void)addCommodityInShoppingCarAddOneOrReduceOne:(CommodityModel *)item {
-    //1.首先到t_shoppingcar中查询，看是否已存在该item
-    //得到结果集
-    NSLog(@"更新item = %@\n",item);
-    if(item.barcode == nil)
-        return;
++ (BOOL)addCommodityInShoppingCarAddOneOrReduceOne:(CommodityModel *)item {
+    
+    if(item == nil||item.barcode == nil)
+        return false;
     FMResultSet * set = [_db executeQueryWithFormat:@"SELECT * FROM t_shoppingcar WHERE barcode = %@",item.barcode];
     NSInteger categoryId;
     NSString * promotionType = @"";
@@ -114,21 +76,19 @@ static FMDatabase * _db;
             promotionType = @"salesAll";
         }
     }
-    NSLog(@"增加或删除一个");
     if (set.next) {
-        NSLog(@"更新购物车商品信息");
-//        [_db executeQueryWithFormat:@"DELETE FROM t_shoppingcar WHERE barcode = %@",item.barcode];
         categoryId = [CommodityManageTool queryCategoryTypeWithCategoryName:item.category];
-        
         [_db executeUpdateWithFormat:@"UPDATE t_shoppingcar SET barcode = %@,name = %@,unit = %@,categoryid = %ld,category = %@,subcategory = %@,price = %f,discountype = %@,count = %ld WHERE barcode = %@", item.barcode,item.name,item.unit,(long)categoryId,item.category,item.subCategory,item.price,promotionType,item.count , item.barcode];
     }
     [_db executeUpdateWithFormat:@"update t_category set isExistInShoppingCar = %d WHERE category = %@",1,item.category];
+    return true;
 }
 
-+ (void)addCommodityInShoppingCar:(CommodityModel *)item {
-    //1.首先到t_shoppingcar中查询，看是否已存在该item
-    NSLog(@"item = %@\n",item);
-    //得到结果集
++ (BOOL)addCommodityInShoppingCar:(CommodityModel *)item
+{
+    if (item == nil) {
+        return false;
+    }
     FMResultSet * set = [_db executeQueryWithFormat:@"SELECT * FROM t_shoppingcar WHERE barcode = %@",item.barcode];
     NSInteger categoryId;
     NSString * promotionType = @"";
@@ -138,39 +98,35 @@ static FMDatabase * _db;
             promotionType = @"salesAll";
         }
     }
-    NSLog(@"set.next = %zd",set.next);
     if (set.next) {
-        NSLog(@"更新购物车");
-//        [_db executeQueryWithFormat:@"DELETE FROM t_shoppingcar WHERE barcode = %@",item.barcode];
-
         categoryId = [CommodityManageTool queryCategoryTypeWithCategoryName:item.category];
-        
         NSInteger count = [set intForColumn:@"count"];
-        
         [_db executeUpdateWithFormat:@"UPDATE t_shoppingcar SET barcode = %@,name = %@,unit = %@,categoryid = %ld,category = %@,subcategory = %@,price = %f,discountype = %@,count = %ld WHERE barcode = %@", item.barcode,item.name,item.unit,(long)categoryId,item.category,item.subCategory,item.price,promotionType,item.count + count , item.barcode];
-        
     } else {
-        NSLog(@"新建商品");
         categoryId = [CommodityManageTool queryCategoryTypeWithCategoryName:item.category];
-        
         [_db executeUpdateWithFormat:@"INSERT INTO t_shoppingcar (barcode,name,unit,categoryid,category,subcategory,price,discountype,count) VALUES (%@,%@,%@,%ld,%@,%@,%f,%@,%ld);",item.barcode,item.name,item.unit,(long)categoryId,item.category,item.subCategory,item.price,promotionType,item.count];
     }
     [_db executeUpdateWithFormat:@"update t_category set isExistInShoppingCar = %d WHERE category = %@",1,item.category];
+    return true;
 }
 
-+ (void)deleteCommodityFromShoppingCar:(CommodityModel *)item {
++ (BOOL)deleteCommodityFromShoppingCar:(CommodityModel *)item
+{
+    if (item == nil) {
+        return false;
+    }
     NSString * deleteStr = [NSString stringWithFormat:@"DELETE FROM t_shoppingcar WHERE barcode = '%@'",item.barcode];
-    NSLog(@"deleteStr = %@\n",deleteStr);
-    BOOL result =  [_db executeUpdate:deleteStr];
-    NSLog(@"result = %zd\n",result);
+    [_db executeUpdate:deleteStr];
     FMResultSet * set = [_db executeQueryWithFormat:@"SELECT * FROM t_shoppingcar WHERE category = %@",item.category];
     if (set.next) {
     } else {
         [_db executeUpdateWithFormat:@"update t_category set isExistInShoppingCar = %d WHERE category = %@",0,item.category];
     }
+    return true;
 }
 
-+ (NSInteger)queryCategoryTypeWithCategoryName:(NSString *)categoryName {
++ (NSInteger)queryCategoryTypeWithCategoryName:(NSString *)categoryName
+{
     FMResultSet * set = [_db executeQueryWithFormat:@"SELECT * FROM t_category WHERE category = %@",categoryName];
     NSInteger categoryId = -1;
     while (set.next) {
@@ -181,12 +137,9 @@ static FMDatabase * _db;
 
 + (NSArray *)commoditiesWithCategory:(NSInteger)categoryId
 {
-    // 得到结果集
     FMResultSet * set = [_db executeQueryWithFormat:@"SELECT * FROM t_commodity_list WHERE categoryid = %ld",categoryId];
-    // 不断往下取数据
     NSMutableArray * commodities = [NSMutableArray array];
     while (set.next) {
-        // 获得当前所指向的数据
         CommodityModel * commodity = [[CommodityModel alloc] init];
         commodity.barcode = [set stringForColumn:@"barcode"];
         commodity.name = [set stringForColumn:@"name"];
@@ -203,12 +156,9 @@ static FMDatabase * _db;
 
 + (NSArray *)commoditiesWithCategory:(NSInteger)categoryId  promotionType:(NSString *)promotionType
 {
-    //得到结果集
     FMResultSet * set = [_db executeQueryWithFormat:@"SELECT * FROM t_commodity_list WHERE categoryid = %ld and discountype = %@",categoryId,promotionType];
-    //不断往下取数据
     NSMutableArray * commodities = [NSMutableArray array];
     while (set.next) {
-        //获得当前所指向的数据
         CommodityModel * commodity = [[CommodityModel alloc] init];
         commodity.barcode = [set stringForColumn:@"barcode"];
         commodity.name = [set stringForColumn:@"name"];
@@ -225,12 +175,9 @@ static FMDatabase * _db;
 
 + (NSArray *)allSalesCommoditiesWithCategory:(NSInteger)categoryId
 {
-    //得到结果集
     FMResultSet * set = [_db executeQueryWithFormat:@"SELECT * FROM t_commodity_list WHERE categoryid = %ld and  discountype != %@ ",categoryId,@""];
-    //不断往下取数据
     NSMutableArray * commodities = [NSMutableArray array];
     while (set.next) {
-        //获得当前所指向的数据
         CommodityModel * commodity = [[CommodityModel alloc] init];
         commodity.barcode = [set stringForColumn:@"barcode"];
         commodity.name = [set stringForColumn:@"name"];
@@ -248,12 +195,9 @@ static FMDatabase * _db;
 
 + (NSArray *)commoditiesInShoppingCarWithCategory:(NSInteger)categoryId
 {
-    // 得到结果集
     FMResultSet * set = [_db executeQueryWithFormat:@"SELECT * FROM t_shoppingcar WHERE categoryid = %ld ",categoryId];
-    // 不断往下取数据
     NSMutableArray * commodities = [NSMutableArray array];
     while (set.next) {
-        // 获得当前所指向的数据
         CommodityModel * commodity = [[CommodityModel alloc] init];
         commodity.barcode = [set stringForColumn:@"barcode"];
         commodity.name = [set stringForColumn:@"name"];
@@ -269,14 +213,11 @@ static FMDatabase * _db;
     return commodities;
 }
 
-+ (NSArray *)categories {
-    
-    // 得到结果集
++ (NSArray *)categories
+{
     FMResultSet * set = [_db executeQueryWithFormat:@"SELECT * FROM t_category "];
-    // 不断往下取数据
     NSMutableArray * categories = [NSMutableArray array];
     while (set.next) {
-        // 获得当前所指向的数据
         CommodityModel * commodity = [[CommodityModel alloc] init];
         commodity.categoryId = [set intForColumn:@"categoryid"];
         commodity.category = [set stringForColumn:@"category"];
@@ -285,14 +226,11 @@ static FMDatabase * _db;
     return categories;
 }
 
-+ (NSArray *)categoriesInShoppingCar {
-    
-    // 得到结果集
++ (NSArray *)categoriesInShoppingCar
+{
     FMResultSet * set = [_db executeQueryWithFormat:@"SELECT * FROM t_category WHERE isExistInShoppingCar = %d",1];
-    // 不断往下取数据
     NSMutableArray * categories = [NSMutableArray array];
     while (set.next) {
-        // 获得当前所指向的数据
         CommodityModel * commodity = [[CommodityModel alloc] init];
         commodity.categoryId = [set intForColumn:@"categoryid"];
         commodity.category = [set stringForColumn:@"category"];
@@ -303,12 +241,9 @@ static FMDatabase * _db;
 
 + (NSArray *)commoditiesInShoppingCar
 {
-    // 得到结果集
     FMResultSet * set = [_db executeQuery:@"SELECT * FROM t_shoppingcar "];
-    // 不断往下取数据
     NSMutableArray * commodities = [NSMutableArray array];
     while (set.next) {
-        // 获得当前所指向的数据
         CommodityModel * commodity = [[CommodityModel alloc] init];
         commodity.barcode = [set stringForColumn:@"barcode"];
         commodity.name = [set stringForColumn:@"name"];
@@ -324,13 +259,11 @@ static FMDatabase * _db;
     return commodities;
 }
 
-+ (void)clearShoppingCar
++ (BOOL)clearShoppingCar
 {
-    BOOL result =  [_db executeUpdate:@"DELETE FROM t_shoppingcar"];
-    NSLog(@"result = %zd\n",result);
+    BOOL result = [_db executeUpdate:@"DELETE FROM t_shoppingcar"];
     [_db executeUpdateWithFormat:@"update t_category set isExistInShoppingCar = %d ",0];
+    return result;
 }
-/*
- *
- */
+
 @end
